Wow, that's a lot of changes.
The VDC is being reworked. We'll abandon the draw-on-the-fly approach (which is much harder to do right and a lot slower) in favor of a sleek, region-based system. This should perform really well and it's very elegant and straightforward.

This is just initial commit of a series of major changes. The emulator is broken now and will remain broken for the next few commits. Collisions and mid-screen changes aren't implemented yet, and there are big bugs that need to be squashed so that even the most basic games can be playable.